moved parse_input_rdms
